requirements.txt
Django==4.1.*
environs==9.5.*
django-phonenumber-field==7.0.*
phonenumbers==8.12.*
Pillow==9.2.*
python-telegram-bot==13.*
